style: 도메인명 스네이크 케이스 변경
