--- src/main/java/com/hyetaekon/hyetaekon/common/public_data/controller/PublicServiceDataController.java
@@ -1,4 +1,0 @@
-package com.hyetaekon.hyetaekon.common.public_data.controller;
-
-public class PublicServiceDataController {
-}
