--- src/main/java/com/hyetaekon/hyetaekon/common/public_data/dto/PublicServiceConditionsDataParsedResponseDto.java
@@ -1,4 +1,0 @@
-package com.hyetaekon.hyetaekon.common.public_data.dto;
-
-public class PublicServiceConditionsDataParsedResponseDto {
-}
