--- src/main/java/com/hyetaekon/hyetaekon/common/public_data/mapper/PublicServiceDataMapper.java
@@ -1,4 +1,0 @@
-package com.hyetaekon.hyetaekon.common.public_data.mapper;
-
-public interface PublicServiceDataMapper {
-}
