--- src/main/java/com/hyetaekon/hyetaekon/common/public_data/schedule/PublicServiceDataScheduler.java
@@ -1,4 +1,0 @@
-package com.hyetaekon.hyetaekon.common.public_data.schedule;
-
-public class PublicServiceDataScheduler {
-}
