--- src/main/java/com/hyetaekon/hyetaekon/common/public_data/service/PublicServiceDataProviderService.java
@@ -1,4 +1,0 @@
-package com.hyetaekon.hyetaekon.common.public_data.service;
-
-public class PublicServiceDataProviderService {
-}
