--- src/main/java/com/hyetaekon/hyetaekon/common/public_data/util/PublicDataPath.java
@@ -1,14 +1,0 @@
-package com.hyetaekon.hyetaekon.common.public_data.util;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public enum PublicDataPath {
-    SERVICE_LIST("/serviceList"),
-    SERVICE_DETAIL_LIST("/serviceDetailList"),
-    SERVICE_CONDITIONS_LIST("/supportConditionsList");
-
-    private final String path;
-}
