+++ src/main/java/com/hyetaekon/hyetaekon/common/publicdata/controller/PublicServiceDataController.java
@@ -1,0 +1,4 @@
+package com.hyetaekon.hyetaekon.common.publicdata.controller;
+
+public class PublicServiceDataController {
+}
