+++ src/main/java/com/hyetaekon/hyetaekon/common/publicdata/dto/PublicServiceDataParsedResponseDto.java
@@ -1,0 +1,4 @@
+package com.hyetaekon.hyetaekon.common.publicdata.dto;
+
+public class PublicServiceDataParsedResponseDto {
+}
