+++ src/main/java/com/hyetaekon/hyetaekon/common/publicdata/dto/PublicServiceDetailDataParsedResponseDto.java
@@ -1,0 +1,4 @@
+package com.hyetaekon.hyetaekon.common.publicdata.dto;
+
+public class PublicServiceDetailDataParsedResponseDto {
+}
