+++ src/main/java/com/hyetaekon/hyetaekon/common/publicdata/mapper/PublicServiceDataMapper.java
@@ -1,0 +1,4 @@
+package com.hyetaekon.hyetaekon.common.publicdata.mapper;
+
+public interface PublicServiceDataMapper {
+}
