+++ src/main/java/com/hyetaekon/hyetaekon/common/publicdata/schedule/PublicServiceDataScheduler.java
@@ -1,0 +1,4 @@
+package com.hyetaekon.hyetaekon.common.publicdata.schedule;
+
+public class PublicServiceDataScheduler {
+}
