+++ src/main/java/com/hyetaekon/hyetaekon/common/publicdata/service/PublicServiceDataProviderService.java
@@ -1,0 +1,4 @@
+package com.hyetaekon.hyetaekon.common.publicdata.service;
+
+public class PublicServiceDataProviderService {
+}
