+++ src/main/java/com/hyetaekon/hyetaekon/common/publicdata/util/PublicDataConstants.java
@@ -1,0 +1,10 @@
+package com.hyetaekon.hyetaekon.common.publicdata.util;
+
+public class PublicDataConstants {
+    public static final String SUBSIDY_DATA_END_POINT = "http://api.odcloud.kr/api/gov24/v3";
+
+    // scheduler 사용
+    public static final String SYNC_STATUS_UPDATE = "U";
+    public static final String SYNC_STATUS_NEW = "A";
+    public static final String SYNC_STATUS_DELETE = "D";
+}
