--- src/main/java/com/hyetaekon/hyetaekon/public_service/controller/PublicServiceController.java
@@ -1,4 +1,0 @@
-package com.hyetaekon.hyetaekon.public_service.controller;
-
-public class PublicServiceController {
-}
