--- src/main/java/com/hyetaekon/hyetaekon/public_service/dto/PublicServiceDto.java
@@ -1,4 +1,0 @@
-package com.hyetaekon.hyetaekon.public_service.dto;
-
-public class PublicServiceDto {
-}
