--- src/main/java/com/hyetaekon/hyetaekon/public_service/entity/PublicService.java
@@ -1,4 +1,0 @@
-package com.hyetaekon.hyetaekon.public_service.entity;
-
-public class PublicService {
-}
