--- src/main/java/com/hyetaekon/hyetaekon/public_service/mapper/PublicServiceMapper.java
@@ -1,4 +1,0 @@
-package com.hyetaekon.hyetaekon.public_service.mapper;
-
-public interface PublicServiceMapper {
-}
