--- src/main/java/com/hyetaekon/hyetaekon/public_service/repository/PublicServiceRepository.java
@@ -1,4 +1,0 @@
-package com.hyetaekon.hyetaekon.public_service.repository;
-
-public class PublicServiceRepository {
-}
