--- src/main/java/com/hyetaekon/hyetaekon/public_service/service/PublicServiceService.java
@@ -1,4 +1,0 @@
-package com.hyetaekon.hyetaekon.public_service.service;
-
-public class PublicServiceService {
-}
