+++ src/main/java/com/hyetaekon/hyetaekon/publicservice/controller/PublicServiceController.java
@@ -1,0 +1,4 @@
+package com.hyetaekon.hyetaekon.publicservice.controller;
+
+public class PublicServiceController {
+}
