+++ src/main/java/com/hyetaekon/hyetaekon/publicservice/dto/PublicServiceDto.java
@@ -1,0 +1,4 @@
+package com.hyetaekon.hyetaekon.publicservice.dto;
+
+public class PublicServiceDto {
+}
