+++ src/main/java/com/hyetaekon/hyetaekon/publicservice/entity/PublicService.java
@@ -1,0 +1,4 @@
+package com.hyetaekon.hyetaekon.publicservice.entity;
+
+public class PublicService {
+}
