+++ src/main/java/com/hyetaekon/hyetaekon/publicservice/mapper/PublicServiceMapper.java
@@ -1,0 +1,4 @@
+package com.hyetaekon.hyetaekon.publicservice.mapper;
+
+public interface PublicServiceMapper {
+}
