+++ src/main/java/com/hyetaekon/hyetaekon/publicservice/repository/PublicServiceRepository.java
@@ -1,0 +1,4 @@
+package com.hyetaekon.hyetaekon.publicservice.repository;
+
+public class PublicServiceRepository {
+}
