+++ src/main/java/com/hyetaekon/hyetaekon/publicservice/service/PublicServiceService.java
@@ -1,0 +1,4 @@
+package com.hyetaekon.hyetaekon.publicservice.service;
+
+public class PublicServiceService {
+}
